Add submission for 02-margin-and-padding-2
Add styling to three boxes within a div container. Make each box stack
on top of each other and have a small amount of space between each
other. Make the padding of the div container equal to the space between
the boxes. Add other styling to make the interior of each box nice and
neat.